refactor: 산책 추천 기능 패키지 분리 (#39)
- TrackingRecommendationController:
  - 일반 산책 Controller에서 추천 API와 요청값 검증 분리
  - 기존 GET /api/tracking/recommendations 주소 유지
- TrackingRecommendationService:
  - 사용자 좌표와 검색 영역 생성 및 추천 응답 변환 책임 이동
- TrackingRecommendationRepository:
  - 추천 전용 QueryDSL 조회 계약으로 이름과 패키지 변경
- QueryDslTrackingRecommendationRepository:
  - 공간 인덱스 후보 조회와 공개 경로 필터링 구현 이동
- RecommendedRouteResponse:
  - 추천 기능 전용 response 패키지로 이동
- TrackingController, TrackingQueryService, TrackingRouteRepository:
  - 추천 책임을 제거하고 일반 산책 조회와 영속성 책임만 유지
