+++ src/main/java/org/zerock/puppyrun/tracking/controller/TrackingController.java
@@ -1,10 +1,6 @@
 package org.zerock.puppyrun.tracking.controller;
 
 import jakarta.validation.Valid;
-import jakarta.validation.constraints.DecimalMax;
-import jakarta.validation.constraints.DecimalMin;
-import jakarta.validation.constraints.Max;
-import jakarta.validation.constraints.Min;
 import java.util.List;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
@@ -19,14 +15,11 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestPart;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
-import org.springframework.validation.annotation.Validated;
 import org.zerock.puppyrun.common.auth.security.UserPrincipal;
 import org.zerock.puppyrun.tracking.controller.request.RegisterTrackingRequest;
 import org.zerock.puppyrun.tracking.controller.response.MainTrackingResponse;
-import org.zerock.puppyrun.tracking.controller.response.RecommendedRouteResponse;
 import org.zerock.puppyrun.tracking.controller.response.TrackingDetailResponse;
 import org.zerock.puppyrun.tracking.controller.request.ChangeVisibilityRequest;
 import org.zerock.puppyrun.tracking.controller.request.UpdateTrackingRequest;
@@ -34,7 +27,6 @@ import org.zerock.puppyrun.tracking.service.TrackingCommandService;
 import org.zerock.puppyrun.tracking.service.TrackingQueryService;
 
 @RestController
-@Validated
 @RequiredArgsConstructor
 @RequestMapping("/api/tracking")
 public class TrackingController {
@@ -59,42 +51,6 @@ public class TrackingController {
     public ResponseEntity<MainTrackingResponse> getTrackingList(
             @AuthenticationPrincipal UserPrincipal userPrincipal) {
         MainTrackingResponse response = trackingQueryService.getTrackingListResponse(userPrincipal.id());
-        return ResponseEntity.ok(response);
-    }
-
-    /**
-     * 사용자 현재 위치 주변에서 시작하는 공개 산책 경로를 조회합니다.
-     *
-     * @param latitude 사용자 현재 위도(-90~90)
-     * @param longitude 사용자 현재 경도(-180~180)
-     * @param radiusMeters 검색 반경(m), 기본 3km
-     * @param limit 최대 추천 개수, 기본 10개
-     * @return 사용자 위치와 가까운 순서로 정렬된 추천 경로
-     */
-    @GetMapping("/recommendations")
-    public ResponseEntity<RecommendedRouteResponse> getRecommendedRoutes(
-            @RequestParam
-            @DecimalMin(value = "-90.0", message = "위도는 -90 이상이어야 합니다.")
-            @DecimalMax(value = "90.0", message = "위도는 90 이하여야 합니다.")
-            double latitude,
-            @RequestParam
-            @DecimalMin(value = "-180.0", message = "경도는 -180 이상이어야 합니다.")
-            @DecimalMax(value = "180.0", message = "경도는 180 이하여야 합니다.")
-            double longitude,
-            @RequestParam(defaultValue = "3000")
-            @Min(value = 100, message = "검색 반경은 100m 이상이어야 합니다.")
-            @Max(value = 50000, message = "검색 반경은 50km 이하여야 합니다.")
-            int radiusMeters,
-            @RequestParam(defaultValue = "10")
-            @Min(value = 1, message = "추천 개수는 1개 이상이어야 합니다.")
-            @Max(value = 50, message = "추천 개수는 50개 이하여야 합니다.")
-            int limit) {
-        RecommendedRouteResponse response = trackingQueryService.getRecommendedRoutes(
-                latitude,
-                longitude,
-                radiusMeters,
-                limit
-        );
         return ResponseEntity.ok(response);
     }
