+++ src/main/java/org/zerock/puppyrun/tracking/recommendation/controller/response/RecommendedRouteResponse.java
@@ -1,4 +1,4 @@
-package org.zerock.puppyrun.tracking.controller.response;
+package org.zerock.puppyrun.tracking.recommendation.controller.response;
 
 import java.util.List;
 import java.util.UUID;
